Simplify and refactor
